+++ mzLib/Test/ProteomicsTests/ProteolyticDigestion/PeptideEqualityAndCloneTests.cs
@@ -16,16 +16,15 @@
 // You should have received a copy of the GNU Lesser General Public
 // License along with Proteomics. If not, see <http://www.gnu.org/licenses/>.
 
-using Chemistry;
-using NUnit.Framework;
-using Assert = NUnit.Framework.Legacy.ClassicAssert;
-using Proteomics.AminoAcidPolymer;
 using System;
 using System.Collections.Generic;
-using System.Linq;
+using Chemistry;
+using NUnit.Framework;
+using Proteomics.AminoAcidPolymer;
+using Assert = NUnit.Framework.Legacy.ClassicAssert;
 using Stopwatch = System.Diagnostics.Stopwatch;
 
-namespace Test
+namespace Test.ProteomicsTests.ProteolyticDigestion
 {
     [TestFixture]
     [System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
