+++ mzLib/Test/ProteomicsTests/ProteolyticDigestion/PeptideWithSetModsDecoyTests.cs
@@ -1,26 +1,19 @@
-using Chemistry;
-using NUnit.Framework;
-using Assert = NUnit.Framework.Legacy.ClassicAssert;
-using CollectionAssert = NUnit.Framework.Legacy.CollectionAssert;
-using Proteomics;
-using Proteomics.ProteolyticDigestion;
 using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using MzLibUtil;
-using Omics;
-using Omics.BioPolymer;
+using Chemistry;
+using NUnit.Framework;
 using Omics.Digestion;
 using Omics.Fragmentation;
 using Omics.Modifications;
-using Omics.SequenceConversion;
-using Transcriptomics.Digestion;
+using Proteomics;
+using Proteomics.ProteolyticDigestion;
 using UsefulProteomicsDatabases;
+using Assert = NUnit.Framework.Legacy.ClassicAssert;
 using Stopwatch = System.Diagnostics.Stopwatch;
-using Transcriptomics;
 
-namespace Test
+namespace Test.ProteomicsTests.ProteolyticDigestion
 {
     [TestFixture]
     [System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
