+++ mzLib/Test/ProteomicsTests/ProteolyticDigestion/PeptideWithSetModsDigestionTests.cs
@@ -1,26 +1,18 @@
-using Chemistry;
-using NUnit.Framework;
-using Assert = NUnit.Framework.Legacy.ClassicAssert;
-using CollectionAssert = NUnit.Framework.Legacy.CollectionAssert;
-using Proteomics;
-using Proteomics.ProteolyticDigestion;
 using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
-using MzLibUtil;
-using Omics;
+using Chemistry;
+using NUnit.Framework;
 using Omics.BioPolymer;
 using Omics.Digestion;
 using Omics.Fragmentation;
 using Omics.Modifications;
-using Omics.SequenceConversion;
-using Transcriptomics.Digestion;
-using UsefulProteomicsDatabases;
+using Proteomics;
+using Proteomics.ProteolyticDigestion;
+using Assert = NUnit.Framework.Legacy.ClassicAssert;
 using Stopwatch = System.Diagnostics.Stopwatch;
-using Transcriptomics;
 
-namespace Test
+namespace Test.ProteomicsTests.ProteolyticDigestion
 {
     [TestFixture]
     [System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
