+++ mzLib/Test/ProteomicsTests/ProteolyticDigestion/TestSeqCoverage.cs
@@ -9,7 +9,7 @@ using Omics.Digestion;
 using Omics.Modifications;
 using Stopwatch = System.Diagnostics.Stopwatch;
 
-namespace Test
+namespace Test.ProteomicsTests.ProteolyticDigestion
 {
     [TestFixture]
     [System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
@@ -85,7 +85,7 @@ namespace Test
         public static void ReadCustomFile()
         {
             Protein ParentProtein = new Protein("OKAREDY", "accession1");
-            string path = Path.Combine(TestContext.CurrentContext.TestDirectory, "DoubleProtease.tsv");
+            string path = Path.Combine(TestContext.CurrentContext.TestDirectory, "ProteomicsTests", "ProteaseFilesForLoadingTests", "DoubleProtease.tsv");
             Assert.That(File.Exists(path));
 
             var proteaseDict = ProteaseDictionary.LoadProteaseDictionary(path);
